Clean up code and refactor socket assigns
