+++ lib/parzival_web/live/app/profile_live/form_component.ex
@@ -58,11 +58,7 @@ defmodule ParzivalWeb.App.ProfileLive.FormComponent do
     end
   end
 
-  require Logger
-
   defp save_user(socket, :new, user_params) do
-    Logger.info(user_params)
-
     case Accounts.admin_create_user(
            user_params,
            &consume_picture_data(socket, &1),
